Int vs Long cleanup #3 (ruby-core:352)
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@2754 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ error.c
@@ -35,10 +35,10 @@ int ruby_nerrs;
 static void
 err_snprintf(buf, len, fmt, args)
     char *buf, *fmt;
-    int len;
+    long len;
     va_list args;
 {
-    int n;
+    long n;
 
     ruby_set_current_source();
     if (!ruby_sourcefile) {
